+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -46,7 +46,7 @@ public class FilmService {
 
         final Event event = Event.builder()
             .userId(userId)
-            .entityId((long) filmId)
+            .entityId(filmId)
             .eventType(EventType.LIKE)
             .operation(Operation.ADD)
             .build();
@@ -67,7 +67,7 @@ public class FilmService {
 
         final Event event = Event.builder()
             .userId(userId)
-            .entityId((long) filmId)
+            .entityId(filmId)
             .eventType(EventType.LIKE)
             .operation(Operation.REMOVE)
             .build();
